RFID(射频识别技术)/tagbeat-python/requirements.txt
joblib==1.1.0
matplotlib==3.0.3
numpy==1.21.5
pandas==1.3.5
Pillow==9.4.0
scipy==1.7.3
